refactor(pptx): extract shape.rs from lib.rs (verbatim)
Shape/tree/picture/media/table/connector/SmartArt assembly: default_placeholder_transform, is_placeholder, read_cnv_pr, parse_shape, parse_pic_prst_geom, svg_blip_path, parse_picture, png_size_from_bytes, EMU_PER_PX_96DPI, parse_media, parse_table_styles_xml, parse_table, parse_table_row, parse_table_cell, extract_decorative_shapes, parse_sp_tree_node, parse_tx_xfrm, parse_smartart_drawing, offset_slide_element, parse_connector — now pub(crate). shape.rs imports parsers from fill/text/chart, model types + LayoutPlaceholders/PptxZip/TableStyleDef + shared XML helpers from crate, and table_style_presets. lib.rs no longer needs chart::* at root (only its tests do — moved to the test module) nor svg_blip_rid; use shape::*.
